+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableDoubleArray.kt
@@ -224,6 +224,12 @@ public value class ImmutableDoubleArray @PublishedApi internal constructor(
             values.all(predicate)
 
     /**
+     * See [DoubleArray.any]
+     */
+    public inline fun any(predicate: (predicate: Double) -> Boolean): Boolean =
+            values.any(predicate)
+
+    /**
      * Leaves this immutable array as is and returns an ImmutableDoubleArray with all elements
      * sorted according to the natural sort order of the value returned by the [selector].
      *